requirements.txt
# Core Data Science & ML
numpy~=1.26.4
pandas~=2.2.2
scikit-learn~=1.5.0
torch~=2.3.0

# NLP & Transformers
transformers~=4.41.0
sentence-transformers~=3.0.0
sastrawi~=1.0.1
nltk~=3.8.1

# Topic Modeling
bertopic~=0.16.0
umap-learn~=0.5.6
hdbscan~=0.8.33

# API Framework
fastapi~=0.111.0
uvicorn[standard]~=0.29.0
pydantic~=2.7.1

# UI & Visualization
streamlit~=1.35.0
requests~=2.31.0

# External APIs
apify_client~=1.4.0
openai~=1.28.0

# Utilities for Configuration
python-dotenv~=1.0.1
pyyaml~=6.0.1

# LangChain & RAG Ecosystem
langchain~=0.2.1
langgraph~=0.1.0
langchain-community~=0.2.1
langchain-openai~=0.1.7
langchain-huggingface~=0.0.3
faiss-cpu~=1.8.0
pypdf~=4.2.0

# AWS SDK
boto3~=1.34.0

# Database Driver
psycopg2-binary~=2.9.9
rank-bm25
passlib[bcrypt]
pgvector